+++ src/main/kotlin/minesweeper/domain/Block.kt
@@ -1,4 +1,4 @@
-package minesweeper
+package minesweeper.domain
 
 sealed interface Block {
     val blockText: String
