+++ src/main/kotlin/minesweeper/domain/BlockCreator.kt
@@ -1,4 +1,4 @@
-package minesweeper
+package minesweeper.domain
 
 class BlockCreator(
     height: Int,
