+++ src/main/kotlin/minesweeper/domain/Blocks.kt
@@ -1,4 +1,4 @@
-package minesweeper
+package minesweeper.domain
 
 class Blocks(normalBlocks: List<Block>, mineBlocks: List<Block>) {
     val blocks: List<Block> = (normalBlocks + mineBlocks).shuffled()
