+++ src/test/kotlin/minesweeper/BlockTest.kt
@@ -2,6 +2,7 @@ package minesweeper
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
+import minesweeper.domain.Block
 
 class BlockTest : StringSpec({
